backend/pyproject.toml
[project]
name = "ai-hedge-fund"
version = "1.0.0"
description = "An AI-powered hedge fund that uses multiple agents to make trading decisions"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "langchain>=0.3.7",
    "langchain-anthropic==0.3.5",
    "langchain-groq==0.2.3",
    "langchain-openai>=0.3.5",
    "langchain-deepseek>=0.1.2",
    "langchain-ollama==0.3.6",
    "langgraph==0.2.56",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "python-dotenv==1.0.0",
    "matplotlib>=3.9.2",
    "tabulate>=0.9.0",
    "colorama>=0.4.6",
    "questionary>=2.1.0",
    "rich>=13.9.4",
    "langchain-google-genai>=2.0.11",
    "fastapi[standard]>=0.104.0",
    "fastapi-cli>=0.0.7",
    "pydantic>=2.4.2",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.22",
    "alembic>=1.12.0",
    "langchain-gigachat>=0.3.12",
    "langchain-xai>=0.2.5",
    "requests>=2.31.0",
    "supabase>=2.11.0",
    "yfinance>=1.2.0,<2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0"
]

[project.scripts]
backtester = "backend.backtesting.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-arm64", "osx-64", "win-64"]

[tool.pixi.dependencies]
python = ">=3.11"

[tool.pixi.pypi-dependencies]
ai-hedge-fund = { path = ".", editable = true }

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev"], solve-group = "default" }

[tool.pixi.tasks]
start = "uvicorn main:app --host 0.0.0.0 --port 8000"
build-docker = "docker build -t jetikan/runpod-backend:latest ."
push-docker = "docker push jetikan/runpod-backend:latest"
deploy-docker = { depends-on = ["build-docker", "push-docker"] }

[tool.pixi.target.linux-64.tasks]
build-docker = "sudo docker build -t jetikan/runpod-backend:latest ."
push-docker = "sudo docker push jetikan/runpod-backend:latest"

[tool.black]
line-length = 420
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
force_alphabetical_sort_within_sections = true
